conso sheets

+++ routes/api.php
@@ -246,17 +246,19 @@ Route::group(['middleware' => ['jwt.verify']], function () {
   });
 
   Route::prefix('church-member')->controller(ChurchMemberController::class)->group(function () {
-    Route::get('/', 'index');
-    Route::get('/{id}', 'show');
-    Route::post('/', 'create');
-    Route::put('/{id}', 'update');
-    Route::delete('/{id}', 'delete');
-    Route::get('/{id}/tracking-logs', 'trackingLogs')->middleware('permission_org:conso-sheet-index');
+    Route::get('/', 'index')->middleware('permission_org:church-member-index');
+    Route::get('/{id}', 'show')->middleware('permission_org:church-member-index');
+    Route::post('/', 'create')->middleware('permission_org:church-member-create');
+    Route::put('/{id}', 'update')->middleware('permission_org:church-member-update');
+    Route::delete('/{id}', 'delete')->middleware('permission_org:church-member-delete');
+    Route::get('/{id}/tracking-logs', 'trackingLogs')->middleware('permission_org:church-member-index');
     Route::post('/{id}/tracking-logs', 'storeTrackingLog')->middleware('permission_org:conso-sheet-index');
     Route::put('/{id}/tracking-logs/{logId}', 'updateTrackingLog')->middleware('permission_org:conso-sheet-index');
     Route::delete('/{id}/tracking-logs/{logId}', 'deleteTrackingLog')->middleware('permission_org:conso-sheet-index');
     Route::put('/{id}/status', 'updateStatus')->middleware('permission_org:conso-sheet-index');
     Route::get('/{id}/status-logs', 'statusLogs')->middleware('permission_org:conso-sheet-index');
+    Route::get('/{id}/consolidators', 'consolidators')->middleware('permission_org:church-member-consolidator-assign');
+    Route::put('/{id}/consolidators', 'syncConsolidators')->middleware('permission_org:church-member-consolidator-assign');
     Route::get('/{id}/medals', 'medals')->middleware('permission_org:conso-sheet-index');
     Route::post('/{id}/medals', 'storeMedal')->middleware('permission_org:conso-sheet-index');
   });
